+++ pydidas/gui/frames/pyfai_calib_frame.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -24,7 +24,7 @@ to be used within pydidas.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -153,19 +153,7 @@ class PyfaiCalibFrame(BaseFrame):
         BaseFrame.__init__(self, **kwargs)
         self._setup_pyfai_context()
         self._tasks = _create_calib_tasks()
-        self.__export_dialog = PydidasFileDialog(
-            parent=self,
-            dialog_type="save_file",
-            caption="Export experiment context file",
-            formats=DiffractionExperimentIo.get_string_of_formats(),
-            default_extension="yaml",
-            dialog=QtWidgets.QFileDialog.getSaveFileName,
-            qsettings_ref="PyfaiCalibFrame__export",
-            info_string=(
-                "<b>Note:</b> Using yaml format allows to export the mask file name as "
-                "well. pyFAI's poni format does not include the mask file."
-            ),
-        )
+        self.__export_dialog = PydidasFileDialog()
 
     def _setup_pyfai_context(self):
         """
@@ -308,7 +296,17 @@ class PyfaiCalibFrame(BaseFrame):
         """
         Export the poni settings to a file.
         """
-        _fname = self.__export_dialog.get_user_response()
+        _fname = self.__export_dialog.get_saving_filename(
+            caption="Export experiment context file",
+            formats=DiffractionExperimentIo.get_string_of_formats(),
+            default_extension="yaml",
+            dialog=QtWidgets.QFileDialog.getSaveFileName,
+            qsettings_ref="PyfaiCalibFrame__export",
+            info_string=(
+                "<b>Note:</b> Using yaml format allows to export the mask file name as "
+                "well. pyFAI's poni format does not include the mask file."
+            ),
+        )
         if _fname is None:
             return
         _experiment = DiffractionExperiment()